Moved generateInvite to common and fixed parentheses
